Finally, part 2 for Day 19 2022

+++ 2022/19/a.rs
@@ -4,7 +4,7 @@ use common::*;
 
 fn main() {
     const MINUTES: i32 = 24;
-    let mut mem = HashMap::<(Status, i32), i32>::new();
+    let mut mem = HashMap::<(Status, i32), i64>::new();
     let mut i = 1;
     let mut sum = 0;
     for bp in read_blueprints_from_stdin() {
